Update pull_data to have custom handlers for GFW SQL query & the new Data API
